+++ langchain4j-guardrails/src/test/java/dev/langchain4j/guardrails/JsonExtractorOutputGuardrailTests.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.guardrail;
+package dev.langchain4j.guardrails;
 
 import static dev.langchain4j.test.guardrail.GuardrailAssertions.assertThat;
 import static org.mockito.Mockito.any;
@@ -10,6 +10,8 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import dev.langchain4j.data.message.AiMessage;
 import java.util.Map;
 import java.util.stream.Stream;
+import dev.langchain4j.guardrail.GuardrailResult;
+import dev.langchain4j.guardrail.OutputGuardrailResult;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
